Default test2 => test
- Extract ToBeImplementedSuite
